Add viento
Se envia al cliente el viento y el tiempo restante  en cada snapshot

+++ src/server/game/projectiles/server_projectile.h
@@ -29,7 +29,7 @@ private:
     float dragConstant;
     float width;
     float height;
-    float countdown;
+    int countdown;
     bool had_impacted;
 
 public:
